refactor; added PUT /comment endpoint; inital image_links_parser package

+++ models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	Id uint64
-	Name string
+	Id       uint
+	Name     string
 	Password string
 }
